+++ src/Maui/OnScreenSizeMarkup.Maui/Extensions/ValueConversionExtensions.shared.cs
@@ -25,7 +25,7 @@ internal static class ValueConversionExtensions
     {
 	    if (Manager.Current.IsLogEnabled)
 	    {
-			ConsoleHelpers.WriteLine($"Attempting To Convert \"{(value == null ? "null" : value)}\" of type:{(value == null ? "null" : value.GetType())} to Type:{(toType == null ? "null" : toType)} on bindable Property of type:{(bindableProperty == null ? "null" : bindableProperty.ReturnType)}", LogLevels.Verbose);
+			LogHelpers.WriteLine($"Attempting To Convert \"{(value == null ? "null" : value)}\" of type:{(value == null ? "null" : value.GetType())} to Type:{(toType == null ? "null" : toType)} on bindable Property of type:{(bindableProperty == null ? "null" : bindableProperty.ReturnType)}", LogLevels.Verbose);
 	    }
 
 	    if (toType == null)
