+++ src/Maui/OnScreenSizeMarkup.Maui/PlatformDensities/ScreenDensityPlatform.shared.cs
@@ -1,7 +1,10 @@
-namespace OnScreenSizeMarkup.Maui.Handlers;
+using System.Diagnostics.CodeAnalysis;
+
+namespace OnScreenSizeMarkup.Maui.PlatformDensities;
 
 
 #if !(IOS || ANDROID || MACCATALYST || WINDOWS)
+[SuppressMessage("Style", "IDE0040:Adicionar modificadores de acessibilidade")]
 internal static partial class ScreenDensityPlatform
 {
 	public static (double xdpi, double ydpi) GetPixelPerInches()
